Part 2

+++ src/NewMusicApp/Controllers/AlbumsController.cs
@@ -20,9 +20,56 @@ namespace NewMusicApp.Controllers
             _context = context;
         }
         // GET: /<controller>/
-        public IActionResult Index()
+        public IActionResult Index(string searchString, string sortOrder)
         {
-            var albums = _context.Albums.Include(a => a.Artist).Include(a => a.Genre);
+            ViewBag.TitleSort = String.IsNullOrEmpty(sortOrder) ? "title_desc" : "";
+            ViewBag.ArtistSort = sortOrder == "Artist" ? "artist_desc" : "Artist";
+            ViewBag.GenreSort = sortOrder == "Genre" ? "genre_desc" : "Genre";
+            ViewBag.PriceSort = sortOrder == "Price" ? "price_desc" : "Price";
+            ViewBag.LikesSort = sortOrder == "Likes" ? "likes_desc" : "Likes";
+
+            var albums = from m in _context.Albums.Include(a => a.Artist).Include(a => a.Genre)
+                         select m;
+
+            if (!String.IsNullOrEmpty(searchString))
+            {
+                albums = albums.Where(a => a.Title.Contains(searchString) || a.Artist.Name.Contains(searchString) || a.Genre.Name.Contains(searchString));
+            }
+
+            switch (sortOrder)
+            {
+                case "title_desc":
+                    albums = albums.OrderByDescending(a => a.Title);
+                    break;
+                case "Artist":
+                    albums = albums.OrderBy(a => a.Artist.Name);
+                    break;
+                case "artist_desc":
+                    albums = albums.OrderByDescending(a => a.Artist.Name);
+                    break;
+                case "Genre":
+                    albums = albums.OrderBy(a => a.Genre.Name);
+                    break;
+                case "genre_desc":
+                    albums = albums.OrderByDescending(a => a.Genre.Name);
+                    break;
+                case "Price":
+                    albums = albums.OrderBy(a => a.Price);
+                    break;
+                case "price_desc":
+                    albums = albums.OrderByDescending(a => a.Price);
+                    break;
+                case "Likes":
+                    albums = albums.OrderBy(a => a.Likes);
+                    break;
+                case "likes_desc":
+                    albums = albums.OrderByDescending(a => a.Likes);
+                    break;
+                default:
+                    albums = albums.OrderBy(a => a.Title);
+                    break;
+            }
+
             return View(albums.ToList());
         }
 
@@ -34,60 +81,68 @@ namespace NewMusicApp.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(Album album)
+        public IActionResult Create(Album album, string artistName, string genreName)
         {
             ViewBag.Genre = new SelectList(_context.Genres, "GenreID", "Name");
             ViewBag.Artists = new SelectList(_context.Artists, "ArtistID", "Name");
 
+            if (artistName != null)
+            {
+                Artist artist = new Models.Artist();
+                artist.Name = artistName;
+
+                var artistCheck = _context.Artists.Any(a => a.Name == artist.Name);
+                if (artistCheck == false)
+                {
+                    _context.Artists.Add(artist);
+                    _context.SaveChanges();
+                    var art = _context.Artists.Single(a => a.Name == artistName);
+                    album.ArtistID = art.ArtistID;
+                }
+                else
+                {
+                    var art = _context.Artists.Single(a => a.Name == artistName);
+                    album.ArtistID = art.ArtistID;
+                }
+            }
+
+            if (genreName != null)
+            {
+                Genre genre = new Models.Genre();
+                genre.Name = genreName;
+
+                var genreCheck = _context.Genres.Any(g => g.Name == genre.Name);
+                if (genreCheck == false)
+                {
+                    _context.Genres.Add(genre);
+                    _context.SaveChanges();
+                    var gen = _context.Genres.Single(g => g.Name == genreName);
+                    album.GenreID = gen.GenreID;
+
+                }
+                else
+                {
+                    var gen = _context.Genres.Single(g => g.Name == genreName);
+                    album.GenreID = gen.GenreID;
+                }
+                ModelState.Remove("ArtistID");
+                ModelState.Remove("GenreID");
+
+            }
             if (ModelState.IsValid)
             {   
                 _context.Albums.Add(album);
                 _context.SaveChanges();
                 return RedirectToAction("Index");
             }
-
-            return View();
+            //else
+            //{
+            //    _context.Albums.Add(album);
+            //    _context.SaveChanges();
+            //    return RedirectToAction("Index");
+            //}
+            return View(album);
         }
-
-        //public IActionResult Edit(int? id)
-        //{
-        //    if (id == null)
-        //    {
-        //        return NotFound();
-        //    }
-        //    var album = _context.Albums.AsNoTracking().Single(a => a.AlbumID == id);
-        //    if (album == null)
-        //    {
-        //        return NotFound();
-        //    }
-        //    PopulateArtistsDropDownList(album.ArtistID);
-        //    PopulateGenresDropDownList(album.GenreID);
-        //    return View(album);
-        //}
-
-        //[HttpPost, ActionName("Edit")]
-        //public IActionResult EditPost(int? id)
-        //{
-        //    if(id == null)
-        //    {
-        //        return NotFound();
-        //    }
-        //    var album = _context.Albums.Single(a => a.AlbumID == id);
-        //    //if (id != album.AlbumID)
-        //    //{
-        //    //    return RedirectToAction("Index");
-        //    //}
-        //    if (ModelState.IsValid)
-        //    {
-        //        _context.Albums.Update(album);
-        //        _context.SaveChanges();
-
-        //        return RedirectToAction("Index");
-        //    }
-        //    PopulateArtistsDropDownList(album.ArtistID);
-        //    PopulateGenresDropDownList(album.GenreID);
-        //    return View(album);
-        //}
 
         public IActionResult Details(int? id)
         {
@@ -101,6 +156,8 @@ namespace NewMusicApp.Controllers
                 .Include(a => a.Genre)
                 .AsNoTracking()
                 .Single(m => m.AlbumID == id);
+
+            ViewBag.AlbumTitle = album.Title;
 
             if (album == null)
             {
@@ -118,6 +175,7 @@ namespace NewMusicApp.Controllers
             }
 
             var album = _context.Albums.Include(a => a.Artist).Include(a => a.Genre).AsNoTracking().Single(m => m.AlbumID == id);
+            ViewBag.AlbumTitle = album.Title;
             if (album == null)
             {
                 return NotFound();
@@ -144,13 +202,47 @@ namespace NewMusicApp.Controllers
         {
             if (id == null)
             {
+                return NotFound();
+            }
+
+            if (ModelState.IsValid)
+            {
+                var album = _context.Albums.Include(a => a.Artist).Include(a => a.Genre).AsNoTracking().Single(m => m.AlbumID == id);
+                _context.Albums.Remove(album);
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            return View();
+        }
+
+        //[HttpPost, ActionName("Delete")]
+        //public IActionResult DeleteConfirm(int id)
+        //{
+        //    if (ModelState.IsValid)
+        //    {
+        //        var album = _context.Albums.Include(a => a.Artist).Include(a => a.Genre).AsNoTracking().Single(m => m.AlbumID == id);
+        //        _context.Albums.Remove(album);
+        //        _context.SaveChanges();
+        //        return RedirectToAction("Index");
+        //    }
+        //    return View();
+
+        //}
+
+        public IActionResult Likes(int? id)
+        {
+            if(id == null)
+            {
                 return RedirectToAction("Index");
             }
 
-            Album album = _context.Albums.Single(m => m.AlbumID == id);
-            _context.Albums.Remove(album);
-            _context.SaveChanges();
-
+            var album = _context.Albums.Include(a => a.Artist).Include(a => a.Genre).AsNoTracking().Single(m => m.AlbumID == id);
+            if (ModelState.IsValid)
+            {
+                album.Likes++;
+                _context.Entry(album).State = EntityState.Modified;
+                _context.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
     }
